cleaned_project/requirements.txt
boto3>=1.28.0
python-wordpress-xmlrpc>=2.3
requests>=2.31.0
pillow>=10.0.0
